requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.1
huggingface-hub==0.16.4
llama-cpp-python==0.2.82
